requirements.txt
grpcio
protobuf
flask 
flask_cors 